+++ CHANGELOG.md
@@ -1,1 +1,3 @@
 # v2.7.48
+
+* added __bool__ to `SortedList`