+++ CHANGELOG.md
@@ -1,3 +1,5 @@
 # v2.7.48
 
 * added __bool__ to `SortedList`
+* added `queue_get`
+* added __bool__ to `Multirun`