+++ CHANGELOG.md
@@ -1,1 +1,6 @@
-# v2.7.49
+# v2.8
+
+**API breaking change!**
+
+* add `chain_functions`
+* `chain` renamed to `chain_functions`